compile messages
