+++ inlang/source-code/paraglide/paraglide-js/src/compilerV2/inputsType.ts
@@ -1,10 +1,10 @@
 import { escapeForSingleQuoteString } from "../services/codegen/escape.js"
 import { isValidJSIdentifier } from "../services/valid-js-identifier/index.js"
 
-export type Params = Record<string, "NonNullable<unknown>">
+export type InputTypeMap = Record<string, "NonNullable<unknown>">
 
 /**
- * Generates a JSDoc comment from a params object.
+ * Generates a JSDoc comment for an input object
  *
  * Returns an empty string if the params object is empty.
  *
@@ -12,9 +12,9 @@ export type Params = Record<string, "NonNullable<unknown>">
  *   const jsdoc = jsdocFromParams({ name: "NonNullable<unknown>", count: "NonNullable<unknown>" })
  *   const message = `/** ${paramsType} *\/ const mes2 => \`Hello ${params.name}! You have ${params.count} messages.\``
  */
-export const paramsType = (params: Params, isMessagesIndex: boolean) => {
+export const inputsType = (params: InputTypeMap, isMessagesIndex: boolean) => {
 	if (Object.keys(params).length === 0) {
-		return isMessagesIndex ? "@param {{}} params" : ""
+		return isMessagesIndex ? "@param {{}} inputs" : ""
 	}
 
 	const fieldTypes: `${string}: ${string}`[] = []
@@ -26,5 +26,5 @@ export const paramsType = (params: Params, isMessagesIndex: boolean) => {
 		}
 	}
 
-	return `@param {{ ${fieldTypes.join(", ")} }} params`
+	return `@param {{ ${fieldTypes.join(", ")} }} inputs`
 }
