+++ inlang/source-code/paraglide/paraglide-js/src/services/codegen/quotes.ts
@@ -1,3 +1,6 @@
+/** Wrap the given string in single quotes */
 export const singleQuote = (str: string) => `'${str}'`
+/** Wrap the given string in double quotes */
 export const doubleQuote = (str: string) => `"${str}"`
+/** Wrap the given string in backticks */
 export const backtick = (str: string) => `\`${str}\``
